Added 2 benchmarks Added result.md file with result output Configure go.mod Adde Makefile Move logic from core to packet
