+++ go.mod
@@ -1,0 +1,3 @@
+module tasks
+
+go 1.15
