--- task2/go.mod
@@ -1,3 +1,0 @@
-module task2
-
-go 1.15
