+++ task3/result.md
@@ -1,0 +1,23 @@
+#### Some results
+
+```json
+[
+    were: 92
+    that: 126
+    from: 68
+    with: 128
+    scarlett: 146 
+    about: 59 
+    said: 200
+    going: 51
+    melanie: 63 
+    went: 62
+]
+``` 
+
+#### Benchmarks
+
+```json
+BenchmarkInsert-4                4746104               262 ns/op
+BenchmarkGetFrequentUses-4      12489802               100 ns/op
+```
